Use the Common project Move Xcode stuff to MacOSProcessManager
