+++ src/Microsoft.DotNet.XHarness.iOS.Shared/Execution/Mlaunch/IMLaunchProcessManager.cs
@@ -10,19 +10,16 @@ using System.Threading.Tasks;
 using Microsoft.DotNet.XHarness.Common.Execution;
 using Microsoft.DotNet.XHarness.Common.Logging;
 
+#nullable enable
 namespace Microsoft.DotNet.XHarness.iOS.Shared.Execution.Mlaunch
 {
-    public interface IMLaunchProcessManager : IProcessManager
+    public interface IMLaunchProcessManager : IMacOSProcessManager, IProcessManager
     {
-
-        string XcodeRoot { get; }
         string MlaunchPath { get; }
-        public Version XcodeVersion { get; }
 
-        Task<ProcessExecutionResult> ExecuteCommandAsync(MlaunchArguments args, ILog log, TimeSpan timeout, Dictionary<string, string> environmentVariables = null, CancellationToken? cancellationToken = null);
-        Task<ProcessExecutionResult> ExecuteCommandAsync(MlaunchArguments args, ILog log, ILog stdoutLog, ILog stderrLog, TimeSpan timeout, Dictionary<string, string> environmentVariables = null, CancellationToken? cancellationToken = null);
-        Task<ProcessExecutionResult> RunAsync(Process process, MlaunchArguments args, ILog log, TimeSpan? timeout = null, Dictionary<string, string> environmentVariables = null, CancellationToken? cancellationToken = null, bool? diagnostics = null);
-        Task<ProcessExecutionResult> RunAsync(Process process, MlaunchArguments args, ILog log, ILog stdoutLog, ILog stderrLog, TimeSpan? timeout = null, Dictionary<string, string> environmentVariables = null, CancellationToken? cancellationToken = null, bool? diagnostics = null);
-        Task<ProcessExecutionResult> ExecuteXcodeCommandAsync(string executable, IList<string> args, ILog log, TimeSpan timeout);
+        Task<ProcessExecutionResult> ExecuteCommandAsync(MlaunchArguments args, ILog log, TimeSpan timeout, Dictionary<string, string>? environmentVariables = null, CancellationToken? cancellationToken = null);
+        Task<ProcessExecutionResult> ExecuteCommandAsync(MlaunchArguments args, ILog log, ILog stdoutLog, ILog stderrLog, TimeSpan timeout, Dictionary<string, string>? environmentVariables = null, CancellationToken? cancellationToken = null);
+        Task<ProcessExecutionResult> RunAsync(Process process, MlaunchArguments args, ILog log, TimeSpan? timeout = null, Dictionary<string, string>? environmentVariables = null, CancellationToken? cancellationToken = null, bool? diagnostics = null);
+        Task<ProcessExecutionResult> RunAsync(Process process, MlaunchArguments args, ILog log, ILog stdoutLog, ILog stderrLog, TimeSpan? timeout = null, Dictionary<string, string>? environmentVariables = null, CancellationToken? cancellationToken = null, bool? diagnostics = null);
     }
 }
