+++ tools/Microsoft.DotNet.XHarness.SimulatorInstaller/Commands/HelpCommand.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.DotNet.XHarness.CLI;
+using Microsoft.DotNet.XHarness.Common.CLI;
 using Microsoft.DotNet.XHarness.SimulatorInstaller.Commands;
 using Mono.Options;
