+++ tools/Microsoft.DotNet.XHarness.SimulatorInstaller/Microsoft.DotNet.XHarness.SimulatorInstaller.csproj
@@ -22,12 +22,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Microsoft.DotNet.XHarness.iOS.Shared\Microsoft.DotNet.XHarness.iOS.Shared.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Include="..\..\src\Microsoft.DotNet.XHarness.CLI\ExitCode.cs" Link="Microsoft.DotNet.XHarness.CLI\ExitCode.cs" />
-    <Compile Include="..\..\src\Microsoft.DotNet.XHarness.CLI\Commands\XHarnessCommand.cs" Link="Microsoft.DotNet.XHarness.CLI\Commands\XHarnessCommand.cs" />
-    <Compile Include="..\..\src\Microsoft.DotNet.XHarness.CLI\CommandArguments\XHarnessCommandArguments.cs" Link="Microsoft.DotNet.XHarness.CLI\CommandArguments\XHarnessCommandArguments.cs" />
+    <ProjectReference Include="..\..\src\Microsoft.DotNet.XHarness.Common\Microsoft.DotNet.XHarness.Common.csproj" />
   </ItemGroup>
 
 </Project>
